+++ src/const.js
@@ -23,6 +23,9 @@ export const effect_jinx_immunity = "Compendium.pf2e-action-support.action-suppo
 export const effect_jinx_clumsy1 = "Compendium.pf2e-action-support.action-support.Item.aEhh2kWTcKJacdHe";
 export const effect_jinx_clumsy2 = "Compendium.pf2e-action-support.action-support.Item.MkWyAtPU3CAFpgdQ";
 export const effect_aberrant_whispers_immunity = "Compendium.pf2e-action-support.action-support.Item.zt5GIv9435SF0fYn";
+export const effect_intimidating_strike = "Compendium.pf2e-action-support.action-support.Item.w9i0aY2IQ3jvCX9K";
+export const effect_reach_spell = "Compendium.pf2e-action-support.action-support.Item.jLxv3W9FgO2SbxVC";
+export const effect_spectral_hand = "Compendium.pf2e-action-support.action-support.Item.wArBvC6i6Das3r8I";
 
 export const spellEffectMap = {
     'adaptive-ablation': ['Compendium.pf2e.spell-effects.Item.6GAztnHuQSwAp1k1'],